Create quiz.js
Quizify is an interactive quiz application designed to engage users in fun and educational quizzes. Users begin by providing their names and are then presented with a series of questions on JavaScript. As they answer questions, Quizify calculates their score based on correctness and provides them with their position compared to other participants.